Improved kerning diff.
Kerning diff drops TTXFont. This approach was a remnant from nototools.
The new approach involves parsing the GPOS table directly. Its alot
cleaner and significantly faster.

Please note: The aim in the next few commits is to remove the TTXFont
object. This pr unfortunately removes the checks which still use this
object.